Add introtowebgl workshop
Add the `introtowebgl` workshop as a dependency.  Since there’s no
validation, add an empty `program.js` to use to mark the exercise
complete.  Also include the `app.js` from the assets folder of the
intro to WebGL workshop with solutions to the first three exercises.

Include a WALKTHROUGH markdown file to possibly use during an Austin
NodeSchool session.